LPS-79038 move from tag to keyword
