+++ modules/apps/headless-apio/blog/blog-apio-impl/build.gradle
@@ -11,9 +11,9 @@ dependencies {
 	compileOnly project(":apps:headless-apio:blog:blog-apio-api")
 	compileOnly project(":apps:headless-apio:category:category-apio-api")
 	compileOnly project(":apps:headless-apio:comment:comment-apio-api")
+	compileOnly project(":apps:headless-apio:keyword:keyword-apio-api")
 	compileOnly project(":apps:headless-apio:media-object:media-object-apio-api")
 	compileOnly project(":apps:headless-apio:person:person-apio-api")
 	compileOnly project(":apps:headless-apio:portal:portal-apio-api")
 	compileOnly project(":apps:headless-apio:site:site-apio-api")
-	compileOnly project(":apps:headless-apio:tag:tag-apio-api")
 }
