+++ modules/apps/headless-apio/keyword/keyword-apio-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Keyword Apio API
+Bundle-SymbolicName: com.liferay.keyword.apio.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.keyword.apio.architect.identifier
