+++ modules/apps/headless-apio/keyword/keyword-apio-api/src/main/resources/com/liferay/keyword/apio/architect/identifier/packageinfo
