+++ modules/apps/headless-apio/keyword/keyword-apio-impl/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Keyword Apio Implementation
+Bundle-SymbolicName: com.liferay.keyword.apio.impl
+Bundle-Version: 1.0.0
+-dsannotations-options: inherit
