+++ modules/apps/headless-apio/keyword/keyword-apio-impl/build.gradle
@@ -2,13 +2,12 @@ sourceCompatibility = "1.8"
 targetCompatibility = "1.8"
 
 dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.1"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly project(":apps:apio-architect:apio-architect-api")
-	compileOnly project(":apps:document-library:document-library-api")
-	compileOnly project(":apps:headless-apio:category:category-apio-api")
-	compileOnly project(":apps:headless-apio:folder:folder-apio-api")
+	compileOnly project(":apps:blogs:blogs-api")
+	compileOnly project(":apps:headless-apio:blog:blog-apio-api")
 	compileOnly project(":apps:headless-apio:keyword:keyword-apio-api")
 	compileOnly project(":apps:headless-apio:media-object:media-object-apio-api")
 	compileOnly project(":apps:headless-apio:person:person-apio-api")
