--- modules/apps/headless-apio/tag/tag-apio-api/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Tag Apio API
-Bundle-SymbolicName: com.liferay.tag.apio.api
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.tag.apio.architect.identifier
