--- modules/apps/headless-apio/tag/tag-apio-impl/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Tag Apio Implementation
-Bundle-SymbolicName: com.liferay.tag.apio.impl
-Bundle-Version: 1.0.0
--dsannotations-options: inherit
