--- modules/apps/headless-apio/tag/tag-apio-impl/build.gradle
@@ -1,16 +1,0 @@
-sourceCompatibility = "1.8"
-targetCompatibility = "1.8"
-
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	compileOnly project(":apps:apio-architect:apio-architect-api")
-	compileOnly project(":apps:blogs:blogs-api")
-	compileOnly project(":apps:headless-apio:blog:blog-apio-api")
-	compileOnly project(":apps:headless-apio:media-object:media-object-apio-api")
-	compileOnly project(":apps:headless-apio:person:person-apio-api")
-	compileOnly project(":apps:headless-apio:portal:portal-apio-api")
-	compileOnly project(":apps:headless-apio:site:site-apio-api")
-	compileOnly project(":apps:headless-apio:tag:tag-apio-api")
-}
